# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Site:
    properties:
      createdAt:
        type: string
      domain:
        type: string
      id:
        type: integer
      identifier:
        type: string
      isActive:
        type: boolean
      isVerified:
        type: boolean
      metadata:
        $ref: '#/definitions/models.SiteMetaData'
      name:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.SiteMetaData:
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isVerified:
        type: boolean
      metadata:
        $ref: '#/definitions/models.UserMetaData'
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      userType:
        $ref: '#/definitions/models.UserType'
    type: object
  models.UserMetaData:
    type: object
  models.UserType:
    enum:
    - ADMIN
    - PUBLISHER
    - ADVERTISER
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RolePublisher
    - RoleAdvertiser
...
host: localhost:8000
info:
  contact: {}
  description: REST API for AdAli.
  title: AdAli API
  version: "1.0"
paths:
...
  /sites:
    get:
      description: |-
        Get all sites or filter and order them using query parameters.

        Filtering operators:
        :   (equals)
        !:  (not equals)
        >   (greater than)
        >=  (greater than or equal)
        <   (less than)
        <=  (less than or equal)
        ~   (contains, case-insensitive)

        Examples:
        /sites?filter=is_active:true
        /sites?filter=user_id:1
        /sites?filter=name~example
        /sites?order_by=created_at&sort=desc
        /sites?order_by=name&order_by=created_at&sort=asc&sort=desc
      parameters:
      - collectionFormat: csv
        description: Filter expression. Can be repeated.
        in: query
        items:
          type: string
        name: filter
        type: array
      - collectionFormat: csv
        description: Fields to order by. Can be repeated.
        in: query
        items:
          type: string
        name: order_by
        type: array
      - collectionFormat: csv
        description: Sort direction for each order_by field (asc or desc). Can be
          repeated.
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Site'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sites
      tags:
      - Sites
    post:
      consumes:
      - application/json
      description: Create a new site.
      parameters:
      - description: Site
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/models.Site'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Site'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create site
      tags:
      - Sites
  /sites/{id}:
    delete:
      description: Delete a site by ID.
      parameters:
      - description: Site ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete site
      tags:
      - Sites
    put:
      consumes:
      - application/json
      description: Update an existing site.
      parameters:
      - description: Site ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated site
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/models.Site'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Site'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update site
      tags:
      - Sites
...
  /users:
    get:
      description: |-
        Get all users or filter and order them using query parameters.

        Filtering operators:
        :   (equals)
        !:  (not equals)
        >   (greater than)
        >=  (greater than or equal)
        <   (less than)
        <=  (less than or equal)
        ~   (contains, case-insensitive)

        Examples:
        /users?filter=is_active:true
        /users?filter=program_id:1
        /users?filter=bid_price>=1000
        /users?order_by=bid_price&sort=desc
        /users?order_by=created_at&order_by=bid_price&sort=desc&sort=asc
      parameters:
      - collectionFormat: csv
        description: Filter expression. Can be repeated.
        in: query
        items:
          type: string
        name: filter
        type: array
      - collectionFormat: csv
        description: Fields to order by. Can be repeated.
        in: query
        items:
          type: string
        name: order_by
        type: array
      - collectionFormat: csv
        description: Sort direction for each order_by field (asc or desc). Can be
          repeated.
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an existing user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - Users
...
swagger: "2.0"
...